move users/status/onboarding to users/onboarding
