+++ controllers/userStatus.js
@@ -1,7 +1,6 @@
 const { fetchUser } = require("../models/users");
 const userStatusModel = require("../models/userStatus");
-const { getUserIdBasedOnRoute, filterUsersWithOnboardingState } = require("../utils/userStatus");
-const { userState } = require("../constants/userStatus");
+const { getUserIdBasedOnRoute } = require("../utils/userStatus");
 const { INTERNAL_SERVER_ERROR } = require("../constants/errorMessages");
 
 /**
@@ -144,72 +143,4 @@ const updateAllUserStatus = async (req, res) => {
   }
 };
 
-/**
- * Collects all Users with ONBOARDING state and are present in discord server for more than 31 days
- *
- * @param req {Object} - Express request object
- * @param res {Object} - Express response object
- */
-
-const getUsersWithOnboardingState = async (req, res) => {
-  let minDaysInServer = 31;
-
-  try {
-    const { allUserStatus } = await userStatusModel.getAllUserStatus(req.query);
-    const { minPresenceDays } = req.query;
-
-    if (!allUserStatus.length) {
-      return res.boom.notFound("User status not found");
-    }
-
-    if (parseInt(minPresenceDays)) {
-      minDaysInServer = parseInt(minPresenceDays);
-    }
-
-    const allUsersWithOnboardingState = filterUsersWithOnboardingState(allUserStatus);
-
-    if (!allUsersWithOnboardingState.length) {
-      return res.boom.notFound("No users exist with an 'ONBOARDING' state");
-    }
-
-    const updatedOnboardingUsersWithDate = [];
-
-    for (const user of allUsersWithOnboardingState) {
-      const result = await fetchUser({ userId: user.userId });
-
-      if (result.user.discordJoinedAt) {
-        const userDiscordJoinedDate = new Date(result.user.discordJoinedAt);
-
-        const currentDate = new Date();
-
-        const timeDifferenceInMilliseconds = currentDate.getTime() - userDiscordJoinedDate.getTime();
-
-        const currentAndUserJoinedDateDifference = Math.floor(timeDifferenceInMilliseconds / (1000 * 60 * 60 * 24));
-
-        if (currentAndUserJoinedDateDifference > minDaysInServer) {
-          updatedOnboardingUsersWithDate.push({ ...result.user, currentStatus: userState.ONBOARDING });
-        }
-      }
-    }
-
-    return res.json({
-      message: updatedOnboardingUsersWithDate.length
-        ? `All Users with ONBOARDING state of more than ${minDaysInServer} days found successfully`
-        : `No users exist with ONBOARDING state of more than ${minDaysInServer} days, or the user has not been verified`,
-      totalUsers: updatedOnboardingUsersWithDate.length,
-      allUser: updatedOnboardingUsersWithDate,
-    });
-  } catch (err) {
-    logger.error(`Error while fetching all the User: ${err}`);
-    return res.boom.badImplementation(INTERNAL_SERVER_ERROR);
-  }
-};
-
-module.exports = {
-  deleteUserStatus,
-  getUserStatus,
-  getAllUserStatus,
-  updateUserStatus,
-  updateAllUserStatus,
-  getUsersWithOnboardingState,
-};
+module.exports = { deleteUserStatus, getUserStatus, getAllUserStatus, updateUserStatus, updateAllUserStatus };
