+++ test/integration/users.test.js
@@ -17,6 +17,7 @@ const searchParamValues = require("../fixtures/user/search")();
 const config = require("config");
 const joinData = require("../fixtures/user/join");
 const {
+  userStatusDataForNewUser,
   userStatusDataAfterSignup,
   userStatusDataAfterFillingJoinSection,
 } = require("../fixtures/userStatus/userStatus");
@@ -1174,6 +1175,69 @@ describe("Users", function () {
           }
           expect(res).to.have.status(200);
           expect(res.body.message).to.be.equal("Successfully added the in_discord field to false for all users");
+          return done();
+        });
+    });
+  });
+
+  describe("GET /users/status/onboarding", function () {
+    let userId1 = "";
+    let userId2 = "";
+    let userId3 = "";
+
+    beforeEach(async function () {
+      userId1 = await addUser(userData[userData.length - 1]);
+      userId2 = await addUser(userData[1]);
+      userId3 = await addUser();
+
+      await userStatusModel.updateUserStatus(userId1, userStatusDataAfterFillingJoinSection);
+      await userStatusModel.updateUserStatus(userId2, userStatusDataForNewUser);
+      await userStatusModel.updateUserStatus(userId3, userStatusDataForNewUser);
+    });
+
+    afterEach(async function () {
+      await cleanDb();
+    });
+
+    it("Should get all the users with ONBOARDING state and are present in discord server for more than 31 days(DEFAULT MIN DATE, without query)", function (done) {
+      chai
+        .request(app)
+        .get("/users/onboarding")
+        .set("Cookie", `${cookieName}=${superUserAuthToken}`)
+        .end((err, res) => {
+          if (err) {
+            return done(err);
+          }
+          expect(res).to.have.status(200);
+          expect(res.body).to.be.a("object");
+          expect(res.body.totalUsers).to.be.a("number");
+          expect(res.body.message).to.equal("All Users with ONBOARDING state of more than 31 days found successfully");
+          expect(res.body.allUser).to.be.a("array");
+          res.body.allUser.forEach((status) => {
+            expect(status).to.have.property("discordJoinedAt");
+          });
+          return done();
+        });
+    });
+
+    it("Should get all the user with ONBOARDING state and are present in discord server for more than (given date in query)", function (done) {
+      chai
+        .request(app)
+        .get("/users/onboarding")
+        .query({ minPresenceDays: 30 })
+        .set("Cookie", `${cookieName}=${superUserAuthToken}`)
+        .end((err, res) => {
+          if (err) {
+            return done(err);
+          }
+          expect(res).to.have.status(200);
+          expect(res.body).to.be.a("object");
+          expect(res.body.totalUsers).to.be.a("number");
+          expect(res.body.message).to.equal("All Users with ONBOARDING state of more than 30 days found successfully");
+          expect(res.body.allUser).to.be.a("array");
+          res.body.allUser.forEach((status) => {
+            expect(status).to.have.property("discordJoinedAt");
+          });
           return done();
         });
     });
